# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Twixtor Archive API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /actors:
    get:
      operationId: listActors
      tags: [actors]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: dramaId
          in: query
          schema: { type: integer }
        - name: type
          in: query
          schema:
            type: string
            enum: [drama, solo]
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActorListResponse"
    post:
      operationId: createActor
      tags: [actors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
  /actors/{id}:
    get:
      operationId: getActor
      tags: [actors]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActorDetail"
    patch:
      operationId: updateActor
      tags: [actors]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActorUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
    delete:
      operationId: deleteActor
      tags: [actors]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
...
  /admin/access-codes:
    get:
      operationId: listAccessCodes
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessCode"
    post:
      operationId: createAccessCode
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessCodeCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessCode"
  /admin/access-codes/{id}:
    delete:
      operationId: deleteAccessCode
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, role]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: ["string", "null"] }
        role: { type: string }
        isVerified: { type: boolean }
        createdAt: { type: string }
    Drama:
      type: object
      required: [id, name, category]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        genre: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        posterUrl: { type: ["string", "null"] }
        videoCount: { type: integer }
        createdAt: { type: string }
...
    Actor:
      type: object
      required: [id, name, type]
      properties:
        id: { type: integer }
        name: { type: string }
        photoUrl: { type: ["string", "null"] }
        type: { type: string, enum: [drama, solo] }
        bio: { type: ["string", "null"] }
        videoCount: { type: integer }
        createdAt: { type: string }
    ActorDetail:
      type: object
      required: [id, name, type]
      properties:
        id: { type: integer }
        name: { type: string }
        photoUrl: { type: ["string", "null"] }
        type: { type: string, enum: [drama, solo] }
        bio: { type: ["string", "null"] }
        videoCount: { type: integer }
        createdAt: { type: string }
        dramas:
          type: array
          items:
            $ref: "#/components/schemas/Drama"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
    ActorListResponse:
      type: object
      required: [actors, total, page, limit]
      properties:
        actors:
          type: array
          items:
            $ref: "#/components/schemas/Actor"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    ActorInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        photoUrl: { type: string }
        type: { type: string, enum: [drama, solo] }
        bio: { type: string }
    ActorUpdate:
      type: object
      properties:
        name: { type: string }
        photoUrl: { type: string }
        type: { type: string, enum: [drama, solo] }
        bio: { type: string }
    Video:
      type: object
      required: [id, title, type, status]
      properties:
        id: { type: integer }
        title: { type: string }
        dramaId: { type: ["integer", "null"] }
        actorId: { type: ["integer", "null"] }
        dramaName: { type: ["string", "null"] }
        actorName: { type: ["string", "null"] }
        episode: { type: ["string", "null"] }
        scene: { type: ["string", "null"] }
        videoUrl: { type: ["string", "null"] }
        thumbnailUrl: { type: ["string", "null"] }
        type: { type: string }
        status: { type: string }
        resolution: { type: ["string", "null"] }
        fps: { type: ["integer", "null"] }
        duration: { type: ["integer", "null"] }
        fileSize: { type: ["integer", "null"] }
        format: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        viewCount: { type: integer }
        downloadCount: { type: integer }
        favoriteCount: { type: integer }
        popularityScore: { type: ["number", "null"] }
        isFavorited: { type: ["boolean", "null"] }
        isBookmarked: { type: ["boolean", "null"] }
        createdAt: { type: string }
...
    UserListResponse:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total: { type: integer }
...
    AccessCode:
      type: object
      required: [id, code]
      properties:
        id: { type: integer }
        code: { type: string }
        isUsed: { type: boolean }
        usedBy: { type: ["integer", "null"] }
        createdAt: { type: string }
    AccessCodeCreate:
      type: object
      properties:
        count: { type: integer }
...