# lib/api-zod/src/generated/types/actorDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Twixtor Archive API
 * OpenAPI spec version: 0.1.0
 */
import type { ActorDetailType } from './actorDetailType';
import type { Drama } from './drama';
import type { Video } from './video';

export interface ActorDetail {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  type: ActorDetailType;
  /** @nullable */
  bio?: string | null;
  videoCount?: number;
  createdAt?: string;
  dramas?: Drama[];
  videos?: Video[];
}