# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Twixtor Archive API
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  title: string;
  /** @nullable */
  dramaId?: number | null;
  /** @nullable */
  actorId?: number | null;
  /** @nullable */
  dramaName?: string | null;
  /** @nullable */
  actorName?: string | null;
  /** @nullable */
  episode?: string | null;
  /** @nullable */
  scene?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  type: string;
  status: string;
  /** @nullable */
  resolution?: string | null;
  /** @nullable */
  fps?: number | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  format?: string | null;
  tags?: string[];
  viewCount?: number;
  downloadCount?: number;
  favoriteCount?: number;
  /** @nullable */
  popularityScore?: number | null;
  /** @nullable */
  isFavorited?: boolean | null;
  /** @nullable */
  isBookmarked?: boolean | null;
  createdAt?: string;
}